+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.3.0] - 2024-8-31
+
+### Added
+
+- sprite configuration (x, y, direction, size, etc.) in the panel
+- Monaco Editor error for complex types (tuple and literal)
+
+### Fixed
+
+- Monaco Editor tokenizer for `self.variables`
+
+### Changed
+
+- dependencies update
+- media elements are now smaller
+
 ## [5.2.0] - 2024-8-22
 
 ### Fixed
@@ -778,8 +794,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Reserved words are no longer in Generator's constructor.
 
-[unreleased]: https://github.com/tomas-wrobel/scrap/compare/v5.2.0...HEAD
-[5.2.0]: https://github.com/tomas-wrobel/scrap/releases/tag/v5.1.2
+[unreleased]: https://github.com/tomas-wrobel/scrap/compare/v5.3.0...HEAD
+[5.3.0]: https://github.com/tomas-wrobel/scrap/releases/tag/v5.3.0
+[5.2.0]: https://github.com/tomas-wrobel/scrap/releases/tag/v5.2.0
 [5.1.2]: https://github.com/tomas-wrobel/scrap/releases/tag/v5.1.2
 [5.1.1]: https://github.com/tomas-wrobel/scrap/releases/tag/v5.1.1
 [5.1.0]: https://github.com/tomas-wrobel/scrap/releases/tag/v5.1.0